extract freeform behavior
